MAINT: rename and leave the doc and public function in python
