+++ scipy/signal/__init__.py
@@ -299,7 +299,7 @@ from ._upfirdn import upfirdn
 from .spline import *
 
 from .bsplines import *
-from .morebsplines import *
+from ._bsplines_pythran import *
 from .filter_design import *
 from .fir_filter_design import *
 from .ltisys import *
